moved all math to a separate file
