+++ StereoPlus2/DomainUtils.hpp
@@ -281,137 +281,3 @@ public:
 	}
 };
 
-class Transform {
-	// Trims angle to 360 degrees
-	static float trimAngle(float a) {
-		int b = a;
-		return b % 360 + (a - b);
-	}
-
-	// Trims angle to [-180;180] and converts to radian
-	static float convertToDeltaDegree(float a) {
-		return trimAngle(a) * 3.1415926f * 2 / 360;
-	}
-
-	// Get rotation angle around the axe.
-	// Calculates a simple sum of absolute values of angles from all axes,
-	// converts angle from radians to degrees
-	// and then trims it to 180 degrees.
-	static float getDirectedDeltaAngle(const glm::vec3& rotation) {
-		float angle = 0;
-		// Rotation axe determines the rotating direction 
-		// so the rotation angles should be positive
-		for (auto i = 0; i < 3; i++)
-			angle += abs(rotation[i]);
-		return trimAngle(angle) * 3.1415926f * 2 / 360;
-	}
-
-	// Get rotation angle around the axe.
-	// Calculates a simple sum of absolute values of angles from all axes,
-	// trims angle to 180 degrees and then converts to radians
-	static bool tryGetLocalRotation(const glm::vec3& rotation, glm::vec3& axe, float& angle) {
-		glm::vec3 t[2] = { glm::vec3(), glm::vec3() };
-		float angles[2] = { 0, 0 };
-
-		// Supports rotation around 1,2 axes simultaneously.
-		auto modifiedAxeNumber = 0;
-		for (auto i = 0; i < 3; i++) {
-			assert(modifiedAxeNumber < 3);
-			if (rotation[i] != 0) {
-				t[modifiedAxeNumber][i] = rotation[i] > 0 ? 1 : -1;
-				angles[modifiedAxeNumber] = abs(rotation[i]);
-				modifiedAxeNumber++;
-			}
-		}
-
-		if (modifiedAxeNumber == 0) {
-			axe = glm::vec3();
-			return false;
-		}
-		if (modifiedAxeNumber == 1) {
-			angle = convertToDeltaDegree(angles[0]);
-			axe = t[0];
-			return true;
-		}
-		
-		angle = convertToDeltaDegree(angles[0] + angles[1]);
-		axe = t[0] + t[1];
-		return true;
-	}
-
-	// Convert local rotation to global rotation relative to cross rotation.
-	// 
-	// Shortly, Isolates rotation in cross' local space.
-	// 	   
-	// Based on quaternion intransitivity, applying new local rotation and then applying inverted initial global rotation 
-	// results in new local rotation being converted to global rotation.
-	// r0 * r1Local * r0^-1 = r1Global
-	// 
-	// Basically, moves rotation initial point from (0;0;0;1) to cross global rotation.
-	// Compromises performance since cross new rotation is calculated both here and ourside of this method
-	// but improves readability and maintainability of the code.
-	static glm::quat getIsolatedRotation(const glm::quat& localRotation, const glm::quat crossGlobalRotation) {
-		auto crossNewRotation = crossGlobalRotation * localRotation;
-		auto rIsolated = crossNewRotation * glm::inverse(crossGlobalRotation);
-		return rIsolated;
-	}
-
-public:
-	static void Scale(const glm::vec3& center, const float& oldScale, const float& scale, std::list<PON>& targets) {
-		for (auto target : targets) {
-			target->SetWorldPosition((target->GetWorldPosition() - center) / oldScale * scale + center);
-			for (size_t i = 0; i < target->GetVertices().size(); i++)
-				target->SetVertice(i, (target->GetVertices()[i] - center) / oldScale * scale + center);
-		}
-	}
-	static void Translate(const glm::vec3& transformVector, std::list<PON>& targets, SceneObject* cross) {
-		// Need to calculate average rotation.
-		// https://stackoverflow.com/questions/12374087/average-of-multiple-quaternions/27410865#27410865
-		if (Settings::SpaceMode().Get() == SpaceMode::Local) {
-			auto r = glm::rotate(cross->GetWorldRotation(), transformVector);
-
-			cross->SetWorldPosition(cross->GetWorldPosition() + r);
-			for (auto o : targets) {
-				o->SetWorldPosition(o->GetWorldPosition() + r);
-				for (size_t i = 0; i < o->GetVertices().size(); i++)
-					o->SetVertice(i, o->GetVertices()[i] + r);
-			}
-			return;
-		}
-
-		cross->SetWorldPosition(cross->GetWorldPosition() + transformVector);
-		for (auto o : targets) {
-			o->SetWorldPosition(o->GetWorldPosition() + transformVector);
-			for (size_t i = 0; i < o->GetVertices().size(); i++)
-				o->SetVertice(i, o->GetVertices()[i] + transformVector);
-		}
-	}
-	static void Rotate(const glm::vec3& center, const glm::vec3& rotation, std::list<PON>& targets, SceneObject* cross) {
-		glm::vec3 axe;
-		float angle;
-
-		if (!tryGetLocalRotation(rotation, axe, angle))
-			return;
-
-		auto r = glm::normalize(glm::angleAxis(angle, axe));
-
-		// Convert local rotation to global rotation.
-		// It is required to rotate all vertices since they don't store their own rotation
-		// meaning that their rotation = (0;0;0;1).
-		//
-		// When multiple objects are selected cross' rotation is equal to the first 
-		// object in selection. See TransformTool::OnSelectionChanged
-		if (Settings::SpaceMode().Get() == SpaceMode::Local)
-			r = getIsolatedRotation(r, cross->GetWorldRotation());
-
-		cross->SetLocalRotation(r * cross->GetLocalRotation());
-
-		for (auto& target : targets) {
-			target->SetWorldPosition(glm::rotate(r, target->GetWorldPosition() - center) + center);
-			for (size_t i = 0; i < target->GetVertices().size(); i++)
-				target->SetVertice(i, glm::rotate(r, target->GetVertices()[i] - center) + center);
-
-			target->SetWorldRotation(r * target->GetWorldRotation());
-		}
-	}
-};
