+++ StereoPlus2/StereoPlus2.vcxproj
@@ -163,6 +163,7 @@
     <ClInclude Include="Json.hpp" />
     <ClInclude Include="Key.hpp" />
     <ClInclude Include="Localization.hpp" />
+    <ClInclude Include="Math.hpp" />
     <ClInclude Include="Renderer.hpp" />
     <ClInclude Include="SceneObject.hpp" />
     <ClInclude Include="SettingsLoader.hpp" />
